Support Patterns as Well as Solid Colors
Use the apa102_patterns module to support all available patterns with
extension/retraction support.

Define more colors, sequences, patterns, & change the RGBColor array to
an array of GenericPatterns.

In an effort to reduce current draw and improve color accuracy, reduce
the number of controlled lights. Move them to the end of the 144 LED
strip to make it easier to swing around & observe patterns.