Enable All 144 LEDs with a 1.25A Current Limit
Modify the Makefile so we use all 144 LEDs in a strip while scaling the
colors so the strip never draws more than 1250 milliamps.

This makes it easy to scale your current draw to the maximum draw of
your voltage regulator & will be crucial as we build a blade using two
strips.